+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/token/generator/collection/CollectionSQLTokenGenerator.java
@@ -15,32 +15,34 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rewrite.token.generator;
+package org.apache.shardingsphere.core.rewrite.token.generator.collection;
 
-import com.google.common.base.Optional;
 import org.apache.shardingsphere.core.rewrite.builder.parameter.ParameterBuilder;
 import org.apache.shardingsphere.core.rewrite.statement.RewriteStatement;
+import org.apache.shardingsphere.core.rewrite.token.generator.SQLTokenGenerator;
 import org.apache.shardingsphere.core.rewrite.token.pojo.SQLToken;
 import org.apache.shardingsphere.core.rule.BaseRule;
 
+import java.util.Collection;
+
 /**
- * SQL token generator for optional.
+ * SQL token generator for collection.
  *
  * @author zhangliang
  * @author panjuan
  * 
  * @param <T> type of rule 
  */
-public interface OptionalSQLTokenGenerator<T extends BaseRule> extends SQLTokenGenerator {
+public interface CollectionSQLTokenGenerator<T extends BaseRule> extends SQLTokenGenerator {
     
     /**
-     * Generate SQL token.
+     * Generate SQL tokens.
      * 
      * @param rewriteStatement rewrite statement
      * @param parameterBuilder SQL parameter builder
      * @param rule rule
      * @param isQueryWithCipherColumn is query with cipher column
-     * @return SQL token
+     * @return SQL tokens
      */
-    Optional<? extends SQLToken> generateSQLToken(RewriteStatement rewriteStatement, ParameterBuilder parameterBuilder, T rule, boolean isQueryWithCipherColumn);
+    Collection<? extends SQLToken> generateSQLTokens(RewriteStatement rewriteStatement, ParameterBuilder parameterBuilder, T rule, boolean isQueryWithCipherColumn);
 }
